pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "web-automator"
version = "0.1.0"
description = "LLM-powered web automation agent with Playwright"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
web-automator = "web_automator.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/web_automator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
